Reversion.

+++ Include/jsonifier/Compare.hpp
@@ -94,7 +94,7 @@ namespace jsonifier_internal {
 				simd_type chunk;
 				while (lengthNew >= vectorSize) {
 					chunk = simd_internal::gatherValuesU<simd_type>(data);
-					mask  = static_cast<uint16_t>(simd_internal::opCmpEq(chunk, search_value));
+					mask  = simd_internal::opCmpEq(chunk, search_value);
 					if JSONIFIER_UNLIKELY (mask != 0) {
 						data += simd_internal::postCmpTzcnt(mask);
 						return data;
